Extracting the BaumeisterFile module
